started adding in a cli to interact with the store
