--- GOALS.md
@@ -1,3 +1,0 @@
-phase 1: only track membership changes (up/down) and propogate them across the cluster.
-
-phase 2: add a meta-data layer that custom data can be propagated across the cluster.
