+++ flip_server.lua
@@ -1,4 +1,3 @@
-#!/usr/local/bin/luvit
 -- -*- mode: lua; tab-width: 2; indent-tabs-mode: 1; st-rulers: [70] -*-
 -- vim: ts=4 sw=4 ft=lua noet
 ---------------------------------------------------------------------
