+++ lib/store-scripts/cli_store.lua
@@ -1,0 +1,17 @@
+-- -*- mode: lua; tab-width: 2; indent-tabs-mode: 1; st-rulers: [70] -*-
+-- vim: ts=4 sw=4 ft=lua noet
+---------------------------------------------------------------------
+-- @author Daniel Barney <daniel@pagodabox.com>
+-- @copyright 2014, Pagoda Box, Inc.
+-- @doc
+--
+-- @end
+-- Created :   4 Feb 2015 by Daniel Barney <daniel@pagodabox.com>
+---------------------------------------------------------------------
+
+-- this will eventually be run by the cli when trying to add
+-- something to the store
+
+return function(bucket,id,...)
+	logger:info("I am trying to add something to the cluster",bucket,id,...)
+end
