+++ lib/store-scripts/init.lua
@@ -1,0 +1,12 @@
+-- -*- mode: lua; tab-width: 2; indent-tabs-mode: 1; st-rulers: [70] -*-
+-- vim: ts=4 sw=4 ft=lua noet
+---------------------------------------------------------------------
+-- @author Daniel Barney <daniel@pagodabox.com>
+-- @copyright 2014, Pagoda Box, Inc.
+-- @doc
+--
+-- @end
+-- Created :   4 Feb 2015 by Daniel Barney <daniel@pagodabox.com>
+---------------------------------------------------------------------
+
+-- this is how the store starts?
